To Do List
In this To Do List app, you can easily add, edit, delete or mark as complete your tasks. There are filter buttons too that help you to filter the tasks accordingly. The tasks you added to this To Do List app will be not stored (It is a sample of to do list). The delete button shows whether you can delete that list or not.